Update settings cell if campaign is running
